Refactor model schema package
