+++ aiosnow/models/__init__.py
@@ -1,3 +1,3 @@
 from .attachment import AttachmentModel
 from .table import TableModel
-from ._base.schema import fields, ModelSchema, ModelSchemaMeta, Pluck, BaseField
+from ._schema import fields, ModelSchema, ModelSchemaMeta, Pluck, BaseField
