+++ aiosnow/models/_base/__init__.py
@@ -1,4 +1,3 @@
-from .schema import fields, ModelSchema, ModelSchemaMeta, Pluck, Nested, BaseField
 from .model import BaseModel, BaseModelMeta
 from .table import BaseTableModel
