+++ aiosnow/models/_schema/fields/string.py
