+++ aiosnow/models/_schema/nested.py
