+++ aiosnow/models/_schema/schema.py
